+++ client/src/MobicoopBundle/Resources/assets/js/components/carpool/results/MatchingJourney.vue
@@ -18,7 +18,9 @@
         <v-icon>mdi-close</v-icon>
       </v-btn>
     </v-toolbar>
-
+    <v-card v-if="fraudWarningDisplay">
+      <v-card-text>{{ $t('fraudWarningText.part1') }} <a :href="$t('fraudWarningText.link')">{{ $t('fraudWarningText.textLink') }}</a>{{ $t('fraudWarningText.part2') }}</v-card-text>
+    </v-card>
     <v-stepper
       v-model="step"
       alt-labels
@@ -599,6 +601,10 @@ export default {
       default: false
     },
     profileSummaryRefresh: {
+      type: Boolean,
+      default: false
+    },
+    fraudWarningDisplay: {
       type: Boolean,
       default: false
     }